+++ test-web/src/components/Charts/ChoroplethMap/Helper/getColor.js
