+++ test-web/src/components/Charts/ChoroplethMap/Helper/processMapData.js
