+++ test-web/src/components/DataDetailPage/AcceptModal.js
@@ -148,3 +148,4 @@ const style = {
   p: 4,
   borderRadius: '10px',
 };
+
