+++ test-web/src/components/DataDetailPage/CardComps/MeatImgsCard.js
@@ -388,3 +388,4 @@ const style = {
     marginBottom: '5px',
   },
 };
+
