+++ test-web/src/components/DataDetailPage/CardComps/MeatImgsCardStatic.js
@@ -188,3 +188,4 @@ const style = {
     border: '1px solid black',
   },
 };
+
