+++ test-web/src/components/DataDetailPage/CardComps/QRInfoCard.js
@@ -100,3 +100,4 @@ const QRInfoCard = ({
 };
 
 export default QRInfoCard;
+
