+++ test-web/src/components/DataDetailPage/DataView.js
@@ -12,7 +12,7 @@ import { TextField, Autocomplete } from '@mui/material';
 // import tables
 import RawTable from './TablesComps/RawTable';
 import ProcessedTable from './TablesComps/ProcessedTable';
-import HeatTable from './TablesComps/arbitaryname';
+import HeatTable from './TablesComps/HeatTable';
 import LabTable from './TablesComps/LabTable';
 import ApiTable from './TablesComps/ApiTable';
 // import timezone
