+++ test-web/src/components/DataDetailPage/NavBar.js
@@ -31,3 +31,4 @@ const NavBar = () => {
 };
 
 export default NavBar;
+
