+++ test-web/src/components/DataDetailPage/RejectModal.js
@@ -149,3 +149,4 @@ const style = {
   p: 4,
   borderRadius: '10px',
 };
+
