+++ test-web/src/components/DataDetailPage/RestrictedModal.js
@@ -86,3 +86,4 @@ const style = {
   p: 4,
   borderRadius: '10px',
 };
+
