+++ test-web/src/components/DataDetailPage/StateChangedModal.js
@@ -116,3 +116,4 @@ const style = {
   p: 4,
   borderRadius: '10px',
 };
+
